Remove suffixes from task names in evaluation results
Task names may be appended by a suffix in the form of -1, -2, etc. during the evaluations, resulting in a variety of keys of the evaluations result dictionary. Currently, this leads to a crash, and it also does not really make sense since it is always the same task and not multiple different ones. To prevent this from happening, the evaluations result dictionary is post-processed, and suffixed in the form "-\d+" are removed.